Move shared value and device helpers into their own modules
Splits parseBool and isTouchscreenDevice out of the catch-all utils
helper so the app store no longer depends on it, and marks the
StoredState import in the store as type-only.

This takes helpers/utils.ts out of the circular import group it was
part of.
