+++ src/components/PanelviewItem.vue
@@ -61,7 +61,7 @@ import MAButton from "@/components/Button.vue";
 import EditorialMediaCard from "@/components/discover/EditorialMediaCard.vue";
 import FavouriteButton from "@/components/FavoriteButton.vue";
 import { handleMenuBtnClick } from "@/helpers/media_item_actions";
-import { parseBool } from "@/helpers/utils";
+import { parseBool } from "@/helpers/parse";
 import { ContentType, type MediaItemType } from "@/plugins/api/interfaces";
 import { getBreakpointValue } from "@/plugins/breakpoint";
 import { computed } from "vue";
