+++ src/helpers/device.ts
@@ -1,0 +1,19 @@
+export function isTouchscreenDevice() {
+  // detect if device/browser is touch enabled
+  let result = false;
+  if (window.PointerEvent && "maxTouchPoints" in navigator) {
+    if (navigator.maxTouchPoints > 0) {
+      result = true;
+    }
+  } else {
+    if (
+      window.matchMedia &&
+      window.matchMedia("(any-pointer:coarse)").matches
+    ) {
+      result = true;
+    } else if (window.TouchEvent || "ontouchstart" in window) {
+      result = true;
+    }
+  }
+  return result;
+}
