+++ tests/helpers/utils.test.ts
@@ -7,7 +7,6 @@ import {
   markdownToHtml,
   numberRange,
   paletteFromServer,
-  parseBool,
   rgbToHex,
   sleep,
   truncateString,
@@ -69,30 +68,6 @@ describe("truncateString", () => {
     expect(truncateString(null, 5)).toBe("");
     // @ts-expect-error testing invalid input
     expect(truncateString(undefined, 5)).toBe("");
-  });
-});
-
-describe("parseBool", () => {
-  it("parses boolean values correctly", () => {
-    expect(parseBool(true)).toBe(true);
-    expect(parseBool(false)).toBe(false);
-  });
-
-  it("parses string values correctly", () => {
-    expect(parseBool("true")).toBe(true);
-    expect(parseBool("false")).toBe(false);
-    expect(parseBool("TRUE")).toBe(true);
-    expect(parseBool("FALSE")).toBe(false);
-  });
-
-  it("handles null/undefined", () => {
-    expect(parseBool(null)).toBe(false);
-    expect(parseBool(undefined)).toBe(false);
-  });
-
-  it("handles empty values", () => {
-    expect(parseBool("")).toBe(false);
-    expect(parseBool("0")).toBe(false);
   });
 });
